Use ImportFileComposite in Adobe Stock
