+++ MediaGallerySynchronization/etc/di.xml
@@ -15,6 +15,7 @@
             <argument name="importers" xsi:type="array">
                 <item name="0" xsi:type="object">Magento\MediaGallerySynchronization\Model\ImportMediaAsset</item>
                 <item name="1" xsi:type="object">Magento\MediaGallerySynchronization\Model\ImportImageFileKeywords</item>
+                <item name="3" xsi:type="object">Magento\MediaGallerySynchronization\Model\ImportThumbnails</item>
             </argument>
         </arguments>
     </type>
